db adjusted to bigint

+++ db/migrate/20220322212452_create_microposts.rb
@@ -2,7 +2,7 @@ class CreateMicroposts < ActiveRecord::Migration[7.0]
   def change
     create_table :microposts do |t|
       t.text :content
-      t.integer :user_id
+      t.bigint :user_id
 
       t.timestamps
     end
