+++ db/migrate/20220324201154_create_comments.rb
@@ -2,7 +2,7 @@ class CreateComments < ActiveRecord::Migration[7.0]
   def change
     create_table :comments do |t|
       t.text :body
-      t.integer :user_id
+      t.bigint :user_id
 
       t.timestamps
     end
