+++ db/migrate/20220324203701_add_micropost_id_to_comments.rb
@@ -1,5 +1,5 @@
 class AddMicropostIdToComments < ActiveRecord::Migration[7.0]
   def change
-    add_column :comments, :micropost_id, :integer
+    add_column :comments, :micropost_id, :bigint
   end
 end
